+++ src/world/generators/flat/FlatWorldGenerator.cs
@@ -8,9 +8,9 @@ namespace DiggersLife
 
 		public FlatWorldGenerator(int height = 64) => Height = height;
 
-		public int Generate(VectorInt position) => Generate(position.X, position.Y);
+		public short Generate(VectorInt position) => Generate(position.X, position.Y);
 
-		public int Generate(int x, int y) =>
+		public short Generate(int x, int y) =>
 		(
 			y == 0 ? Blocks.Bedrock :
 			y < Height - 7 ? Blocks.Stone :
